Fix compiler warnings. (#7974)
- Remove unused parameters. There are still many warnings that are not yet
addressed. Currently, the warnings in dmlc-core dominate the error log.
- Remove `distributed` parameter from metric.
- Fixes some warnings about signed comparison.

+++ src/tree/gpu_hist/row_partitioner.cuh
@@ -116,7 +116,7 @@ class RowPartitioner {
     Segment segment = ridx_segments_.at(nidx);  // rows belongs to node nidx
     auto d_ridx = ridx_.CurrentSpan();
     auto d_position = position_.CurrentSpan();
-    if (left_counts_.size() <= nidx) {
+    if (left_counts_.size() <= static_cast<size_t>(nidx)) {
       left_counts_.resize((nidx * 2) + 1);
       thrust::fill(left_counts_.begin(), left_counts_.end(), 0);
     }
